[Fix] #15 merge conflict resolved
