--- Ninano/Ninano/CalendarDetail/WeeklyCalendarCell.swift
@@ -1,14 +1,0 @@
-//
-//  WeeklyCalendarCell.swift
-//  Ninano
-//
-//  Created by Eunbee Kang on 2022/07/20.
-//
-
-import UIKit
-
-final class WeeklyCalendarCell: UICollectionViewCell {
-    @IBOutlet weak var dayNameLabel: UILabel!
-    @IBOutlet weak var dateNumberLabel: UILabel!
-    @IBOutlet weak var dayHighlight: UIView!
-}
