+++ Ninano/Ninano/Views/CalendarTab/DayEventDetailCell.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class DayEventDetailCell: UITableViewCell {
+final class DayEventDetailCell: UITableViewCell {
     @IBOutlet weak var posterImage: UIImageView!
     @IBOutlet weak var eventNameLabel: UILabel!
     @IBOutlet weak var eventPlaceLabel: UILabel!
